+++ backend/Entities/User.cs
@@ -1,7 +1,9 @@
-﻿using System.Text.Json.Serialization;
+﻿using System.ComponentModel.DataAnnotations;
+using System.Text.Json.Serialization;
 
 namespace backend.Entities
 {
+    //base User Model
     public class User
     {
         public int Id { get; set; }
@@ -11,5 +13,50 @@ namespace backend.Entities
 
         [JsonIgnore]
         public string? PasswordHash { get; set; }
+    }
+
+    //user model for registering users
+    public class RegisterRequest
+    {
+        [Required]
+        public string? FirstName { get; set; }
+
+        [Required]
+        public string? LastName { get; set; }
+
+        [Required]
+        public string? Username { get; set; }
+
+        [Required]
+        public string? Password { get; set; }
+    }
+
+    //user model for updating a user's information
+    public class UpdateRequest
+    {
+        public string? FirstName { get; set; }
+        public string? LastName { get; set; }
+        public string? Username { get; set; }
+        public string? Password { get; set; }
+    }
+
+    //user model for authenticating a response
+    public class AuthenticateResponse
+    {
+        public int Id { get; set; }
+        public string? FirstName { get; set; }
+        public string? LastName { get; set; }
+        public string? Username { get; set; }
+        public string? Token { get; set; }
+    }
+
+    //user model for authenticating a request
+    public class AuthenticateRequest
+    {
+        [Required]
+        public string? Username { get; set; }
+
+        [Required]
+        public string? Password { get; set; }
     }
 }
