--- backend/Helpers/SqliteDataContext.cs
@@ -1,17 +1,0 @@
-﻿using Microsoft.EntityFrameworkCore;
-using backend.Entities;
-using Microsoft.Extensions.Configuration;
-
-namespace backend.Helpers
-{
-    public class SqliteDataContext : DataContext
-    {
-        public SqliteDataContext(IConfiguration configuration) : base(configuration) { }
-
-        protected override void OnConfiguring(DbContextOptionsBuilder options)
-        {
-            // connect to sqlite database
-            options.UseSqlite(Configuration.GetConnectionString("WebApiDatabase"));
-        }
-    }
-}
