--- backend/Models/AuthenticateRequest.cs
@@ -1,13 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace backend.Models
-{
-    public class AuthenticateRequest
-    {
-        [Required]
-        public string? Username { get; set; }
-
-        [Required]
-        public string? Password { get; set; }
-    }
-}
