--- backend/Models/AuthenticateResponse.cs
@@ -1,13 +1,0 @@
-﻿using backend.Entities;
-
-namespace backend.Models
-{
-    public class AuthenticateResponse
-    {
-        public int Id { get; set; }
-        public string? FirstName { get; set; }
-        public string? LastName { get; set; }
-        public string? Username { get; set; }
-        public string? Token { get; set; }
-    }
-}
