+++ backend/Models/DataContext.cs
@@ -1,7 +1,7 @@
 ﻿using backend.Entities;
 using Microsoft.EntityFrameworkCore;
 
-namespace backend.Helpers
+namespace backend.Models
 {
     public class DataContext : DbContext
     {
@@ -14,8 +14,8 @@ namespace backend.Helpers
 
         protected override void OnConfiguring(DbContextOptionsBuilder options)
         {
-            // connect to sql server database
-            options.UseSqlServer(Configuration.GetConnectionString("WebApiDatabase"));
+            // connect to sqlite database
+            options.UseSqlite(Configuration.GetConnectionString("WebApiDatabase"));
         }
 
         public DbSet<User> Users { get; set; }
