--- backend/Models/UpdateRequest.cs
@@ -1,10 +1,0 @@
-﻿namespace backend.Models
-{
-    public class UpdateRequest
-    {
-        public string? FirstName { get; set; }
-        public string? LastName { get; set; }
-        public string? Username { get; set; }
-        public string? Password { get; set; }
-    }
-}
