+++ backend/Program.cs
@@ -1,5 +1,6 @@
 using backend.Authorization;
 using backend.Helpers;
+using backend.Models;
 using backend.Services;
 using Microsoft.EntityFrameworkCore;
 
@@ -14,14 +15,16 @@ namespace backend
             var services = builder.Services;
             var env = builder.Environment;
 
-            // use sql server db in production and sqlite db in development
-            if (env.IsProduction())
-                services.AddDbContext<DataContext>();
-            else
-                services.AddDbContext<DataContext, SqliteDataContext>();
+            //registering DatabaseContext
+            //This will allow us to communicate and perform operation to our database through a data context object
+            services.AddDbContext<DataContext>();
 
             services.AddCors();
             services.AddControllers();
+
+            //allow use of swagger for API testing
+            services.AddEndpointsApiExplorer();
+            services.AddSwaggerGen();
 
             // configure automapper with all automapper profiles from this assembly
             services.AddAutoMapper(typeof(Program));
@@ -30,6 +33,11 @@ namespace backend
             services.Configure<AppSettings>(builder.Configuration.GetSection("AppSettings"));
 
             // configure DI for application services
+            /*
+             * Any object/variable of type IJwtUtils will have all its dependecies resolved by .NET
+             * meaning we do not have to create an instance of the class or any classes that it depends on,
+             * dependancy injection handles that logic.
+             */
             services.AddScoped<IJwtUtils, JwtUtils>();
             services.AddScoped<IUserService, UserService>();
 
@@ -42,9 +50,15 @@ namespace backend
                 dataContext.Database.Migrate();
             }
 
+            if(app.Environment.IsDevelopment())
+            {
+                app.UseSwagger();
+                app.UseSwaggerUI();
+            }
+
             // configure HTTP request pipeline
             {
-                // global cors policy
+                //allow calls to our API from any origin
                 app.UseCors(x => x
                     .AllowAnyOrigin()
                     .AllowAnyMethod()
